refactor(compiler): move config loader into `loaders`
